# packages/api-server/src/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: ConfigBuddy CMDB API
  version: 2.0.0
  description: |
    Comprehensive REST API for ConfigBuddy - Enterprise Configuration Management Database (CMDB) platform.

    ConfigBuddy v2.0 features:
    - Graph-based CI relationships with Neo4j
    - 37 connectors (17 TypeScript + 20 JSON-only)
    - Unified credential system with affinity matching
    - Discovery agent architecture with smart routing
    - AI-powered anomaly detection and impact analysis
    - Identity resolution and reconciliation

    ## Authentication

    All API endpoints (except `/api/v1/auth/login`) require authentication via:
    - **JWT Token**: Include `Authorization: Bearer <token>` header
    - **API Key**: Include `X-API-Key: <key>` header

    ## Rate Limiting

    Rate limits apply per authentication token/API key:
    - **Standard**: 1000 requests/hour
    - **Burst**: 100 requests/minute

    Rate limit headers included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  contact:
    name: ConfigBuddy Support
    url: http://localhost:8080
    email: support@configbuddy.io
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: http://staging-api.configbuddy.io/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /relationships:
    get:
      tags: [Relationships]
      summary: List relationships
      description: Retrieve relationships with optional filtering
      operationId: listRelationships
      parameters:
        - name: _type
          in: query
          schema:
            $ref: '#/components/schemas/RelationshipType'
        - name: _from_id
          in: query
          schema:
            type: string
          description: Filter by source CI ID
        - name: _to_id
          in: query
          schema:
            type: string
          description: Filter by target CI ID
        - name: _ci_id
          in: query
          schema:
            type: string
          description: Filter by any CI ID (source or target)
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of relationships
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Relationship'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Relationships]
      summary: Create relationship
      description: Create a relationship between two CIs
      operationId: createRelationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationshipInput'
            examples:
              dependency:
                summary: Application depends on database
                value:
                  _from_id: app-ecommerce-api
                  _to_id: db-postgres-prod
                  _type: DEPENDS_ON
                  properties:
                    connection_string: postgresql://db-postgres-prod:5432/ecommerce
              hosting:
                summary: Server hosts application
                value:
                  _from_id: srv-prod-web-01
                  _to_id: app-ecommerce-api
                  _type: HOSTS
      responses:
        '201':
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relationship'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: One or both CIs not found
  /relationships/{id}:
    delete:
      tags: [Relationships]
      summary: Delete relationship
      description: Remove a relationship between CIs
      operationId: deleteRelationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Relationship ID or query string from_id/to_id/type
      responses:
        '204':
          description: Relationship deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /relationships/type/{type}:
    get:
      tags: [Relationships]
      summary: Get relationships by type
      description: Retrieve all relationships of a specific type
      operationId: getRelationshipsByType
      parameters:
        - name: type
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/RelationshipType'
      responses:
        '200':
          description: Relationships of specified type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Relationship'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # =============================================================================
  # SEARCH
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from /auth/api-keys endpoint
  parameters:
...
    Limit:
      name: _limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: _offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation Error
            message: Invalid CI type provided
            details:
              - field: _type
                error: must be one of [server, virtual-machine, ...]
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or expired JWT token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: CI with ID 'srv-prod-web-01' not found
...
  schemas:
...
    Relationship:
      type: object
      required: [_from_id, _to_id, _type]
      properties:
        _from_id:
          type: string
          description: Source CI ID
        _to_id:
          type: string
          description: Target CI ID
        _type:
          $ref: '#/components/schemas/RelationshipType'
        properties:
          type: object
          additionalProperties: true
          description: Additional relationship metadata
    RelationshipInput:
      type: object
      required: [_from_id, _to_id, _type]
      properties:
        _from_id:
          type: string
        _to_id:
          type: string
        _type:
          $ref: '#/components/schemas/RelationshipType'
        properties:
          type: object
          additionalProperties: true
    RelationshipType:
      type: string
      enum:
        - DEPENDS_ON
        - HOSTS
        - CONNECTS_TO
        - USES
        - OWNED_BY
        - PART_OF
        - LOCATED_IN
        - DEPLOYED_ON
        - BACKED_UP_BY
        - DETECTED_ON
        - AFFECTS
        - INSTALLED_ON
        - CONTAINS
        - ASSIGNED_TO
        - APPLIES_TO
        - MEMBER_OF
        - REQUIRED_BY
        - HOSTED_ON

    # =========================================================================
    # AUTHENTICATION
    # =========================================================================
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type/code
          example: Validation Error
        message:
          type: string
          description: Human-readable error message
          example: Invalid CI type provided
        details:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
              error:
                type: string